+++ resources/views/back/pages/profile.blade.php
@@ -1,5 +1,5 @@
 @extends('back.layouts.pages-layout')
-@section('pageTitle', isset($pageTitle) ? $pageTitle : 'Profile')
+@section('pageTitle', isset($pageTitle) ? $pageTitle : __('Profile'))
 @section('content')
 
 @livewire('author-profile-header')
@@ -14,14 +14,14 @@
                     <polyline points="5 12 3 12 12 3 21 12 19 12" />
                     <path d="M5 12v7a2 2 0 0 0 2 2h10a2 2 0 0 0 2 -2v-7" />
                     <path d="M9 21v-6a2 2 0 0 1 2 -2h2a2 2 0 0 1 2 2v6" />
-                </svg> Personal details</a>
+                </svg> {{ __('Personal details') }}</a>
             </li>
             <li class="nav-item">
                 <a href="#tabs-password" class="nav-link" data-bs-toggle="tab"><svg xmlns="http://www.w3.org/2000/svg" class="icon me-2" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round">
                     <path stroke="none" d="M0 0h24v24H0z" fill="none" />
                     <circle cx="12" cy="7" r="4" />
                     <path d="M6 21v-2a4 4 0 0 1 4 -4h4a4 4 0 0 1 4 4v2" />
-                </svg> Change password</a>
+                </svg> {{ __('Change password') }}</a>
             </li>
             </ul>
         </div>